requirements.txt
greenlet==1.1.2
grpcio==1.45.0
grpcio-tools==1.45.0
importlib-metadata==4.11.3
protobuf==3.19.4
psycopg2==2.9.3
six==1.16.0
SQLAlchemy==1.4.32
typing-extensions==4.1.1
zipp==3.7.0
